refactor: move mixins to mixins.py file
