+++ eox_nelp/api_clients/tests/tests_certificates.py
@@ -11,7 +11,7 @@ from django.utils import timezone
 from mock import Mock, patch
 
 from eox_nelp.api_clients.certificates import ExternalCertificatesApiClient
-from eox_nelp.api_clients.tests import TestBasicAuthAuthenticatorMixin, TestRestApiClientMixin
+from eox_nelp.api_clients.tests.mixins import TestBasicAuthAuthenticatorMixin, TestRestApiClientMixin
 from eox_nelp.signals.utils import generate_reference_id
