+++ eox_nelp/api_clients/tests/tests_futurex.py
@@ -9,7 +9,7 @@ import requests
 from mock import patch
 
 from eox_nelp.api_clients.futurex import FuturexApiClient, FuturexMissingArguments
-from eox_nelp.api_clients.tests import TestOauth2AuthenticatorMixin, TestRestApiClientMixin
+from eox_nelp.api_clients.tests.mixins import TestOauth2AuthenticatorMixin, TestRestApiClientMixin
 
 
 class TestFuturexApiClient(TestRestApiClientMixin, TestOauth2AuthenticatorMixin, unittest.TestCase):
